refactor: moved analytics code to document.js This commit moves google analytics script to document.js and adds a condition to render ads based on if a user is premium.
